requirements.txt
absl-py==0.7.1
alabaster==0.7.12
alembic==1.3.0
allennlp==0.9.0
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.0
astunparse==1.6.3
async-generator==1.10
attrs==17.4.0
Automat==0.6.0
Babel==2.8.0
backcall==0.1.0
bleach==1.5.0
blinker==1.4
blis==0.2.4
cachetools==4.1.1
certifi==2018.1.18
certipy==0.1.3
chardet==3.0.4
click==6.7
cloud-init==19.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
conllu==1.3.1
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.14
dataclasses==0.8
decorator==4.4.1
defusedxml==0.6.0
dgl-cu102==0.5.2
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
dyNET==2.1
editdistance==0.5.3
en-core-web-sm==2.2.5
entrypoints==0.3
faiss==1.5.3
faiss-gpu==1.6.3
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.18.2
gast==0.3.3
gevent==1.4.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.23.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imagesize==1.2.0
incremental==16.10.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10
jsonnet==0.14.0
jsonpatch==1.16
jsonpickle==1.2
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterhub==0.9.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
line-bot-sdk==1.15.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.3
mistune==0.8.4
mock==3.0.5
mwparserfromhell==0.6.dev0
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.5
notebook==6.0.2
numpy==1.16.4
numpydoc==0.9.2
oauthlib==3.1.0
opencv-python==4.1.0.25
opt-einsum==3.3.0
overrides==2.8.0
packaging==20.1
PAM==0.4.2
pamela==1.0.0
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.12.4
ptyprocess==0.6.0
PuLP==2.1
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
pytest==5.3.5
python-apt==1.6.5+ubuntu0.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python-oauth2==1.1.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
qtconsole==4.5.5
regex==2020.1.8
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
responses==0.10.9
rsa==4.6
scipy==1.4.1
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.85
service-identity==16.0.0
six==1.15.0
sklearn-crfsuite==0.3.6
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
thinc==7.0.8
torch==1.7.0
torch-cluster==1.5.7
torch-scatter==2.0.5
torch-sparse==0.6.7
torch-spline-conv==1.2.0
torchvision==0.8.1
tornado==6.0.3
tqdm==4.50.0
traitlets==4.3.3
Twisted==17.9.0
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.3.2
